Configure environment for unit test
Autotests for this project (Blender and ShakeToken) were implemented with use of
waffle lib https://ethereum-waffle.readthedocs.io/ . This lib try
compile all *.sol files inside ./contracts folder. But there is much of other
contracts code in this folder. Some of it have unresolved dependency on
./SafeMath.sol so waffle can't compile all code. That why code of Blender and
ShakeToken was moved to separate folder ./contracts/autotest . And waffle is configured
for us just this folder as source for compile